LPS-95165 Move JMXWhiteboardTest to module and adapt to arquillian test
