--- portal-impl/test/integration/com/liferay/portal/jmx/bundle/jmxwhiteboard/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Description: Test Bundle for JMXWhiteboardTest
-Bundle-Version: 1.0.0
-Private-Package: $(bundle.package).*
--dsannotations: *
